BUG: Signed/unsigned
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=7264
git-svn-id: http://svn.slicer.org/Slicer4/trunk@7264 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkMRMLGridTransformNode.cxx
@@ -238,7 +238,7 @@ void vtkMRMLGridTransformNode::ReadXMLAttributes(const char** atts)
         {
         vals.push_back( val );
         }
-      if (vals.size() !=  num_of_displacement)
+      if ((int)(vals.size()) !=  num_of_displacement)
         {
         vtkErrorMacro( "Incorrect number of origin: expecting " << num_of_displacement << "; got "
                        << vals.size() );
